Move DateTimeUtil into precog project
